--- dashboard/engine/__init__.py
@@ -1,1 +1,0 @@
-# Jobs Engine
