+++ dashboard/jobs_framework/__init__.py
@@ -1,0 +1,1 @@
+# Jobs Framework
